+++ src/frontend/apps/impress/src/features/docs/doc-editor/hook/index.ts
@@ -1,2 +1,3 @@
+export * from './useHeadings';
 export * from './useSaveDoc';
 export * from './useUploadFile';